requirements.txt
langchain
langchain-groq
langchain-core
gradio
requests
python-dotenv
duckduckgo-search
ddgs
